+++ Tests/TestExecutor.cs
@@ -63,6 +63,7 @@ namespace DynamicBox.Quest.Tests
                     ProgressReportingTests.RunAllProgressTests();
                     QuestContextTests.RunAllContextTests();
                     QuestSerializationTests.RunAllSerializationTests();
+                    ServiceImplementationTests.RunAllServiceTests();
                     EventDrivenConditionTests.RunAllEventDrivenTests();
                     FactoryMethodTests.RunAllFactoryMethodTests();
                     ImmutableEventTests.RunAllImmutableEventTests();